Revert "chore: remove additional skeleton component (#558)"
This reverts commit 6b498d25e4bccf3893d3cc3d4f88a54b28e1b172.
